CHANGE: 코드정리
