+++ jisu/appDataSource.js
@@ -1,0 +1,14 @@
+const {DataSource} = require('typeorm');
+
+const appDataSource = new DataSource({
+  type: process.env.DB_TYPE,
+  host: process.env.DB_HOST,
+  port: process.env.DB_PORT,
+  username: process.env.DB_USER,
+  password: process.env.DB_PASSWORD,
+  database: process.env.DB_NAME,
+});
+
+module.exports = {
+    appDataSource
+}
